% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_ALK_KYNTH.R
\name{Calc.ALK.KYNTH}
\alias{Calc.ALK.KYNTH}
\title{XXX}
\usage{
Calc.ALK.KYNTH(OTH.KYNTH, OTH.TOT, ALK.TOT, kfile, aldur, lengd)
}
\arguments{
\item{OTH.KYNTH}{XXX}

\item{OTH.TOT}{XXX}

\item{ALK.TOT}{XXX}

\item{kfile}{XXX}

\item{aldur}{XXX}

\item{lengd}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

