% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_ALK.R
\name{Calc.ALK}
\alias{Calc.ALK}
\title{XXX}
\usage{
Calc.ALK(OTH.TOT, kfile, aldur, lengd)
}
\arguments{
\item{OTH.TOT}{XXX}

\item{kfile}{XXX}

\item{aldur}{XXX}

\item{lengd}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

