% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_fj_per_station.R
\name{Calc.fj.per.station}
\alias{Calc.fj.per.station}
\title{Title}
\usage{
Calc.fj.per.station(ALK, LDIST)
}
\arguments{
\item{ALK}{XXX}

\item{LDIST}{XXX}
}
\value{
XXX
}
\description{
location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

