% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_index.R
\name{Calc.index}
\alias{Calc.index}
\title{Calc.index}
\usage{
Calc.index(sfile, colname, strata.list, std.toglengd = 4, trollbreidd = 17,
  combine.output = Std.aggregation, use.rallarea = T,
  leidretta.fyrir.toglengd = T, z, std.cv = 1, mintoglengd, maxtoglengd)
}
\arguments{
\item{sfile}{XXX}

\item{colname}{XXX}

\item{strata.list}{XXX}

\item{std.toglengd}{XXX}

\item{trollbreidd}{XXX}

\item{combine.output}{XXX}

\item{use.rallarea}{XXX}

\item{leidretta.fyrir.toglengd}{XXX}

\item{z}{XXX}

\item{std.cv}{XXX}

\item{mintoglengd}{XXX}

\item{maxtoglengd}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

