% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcgear.R
\name{calcgear}
\alias{calcgear}
\title{XXX}
\usage{
calcgear(sfile, gearlist = pax::gearlist)
}
\arguments{
\item{sfile}{XXX}

\item{gearlist}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

