% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_alk_visit.R
\name{combine.alk.visit}
\alias{combine.alk.visit}
\title{XXX}
\usage{
combine.alk.visit(fjvisit, kynthvisit, biovisit, meanlevisit, sdevvisit,
  kynthbiovisit, ar = 1985:2002, aldur = 1:11, rows = 22, mult = 1e+06)
}
\arguments{
\item{fjvisit}{XXX}

\item{kynthvisit}{XXX}

\item{biovisit}{XXX}

\item{meanlevisit}{XXX}

\item{sdevvisit}{XXX}

\item{kynthbiovisit}{XXX}

\item{ar}{XXX}

\item{aldur}{XXX}

\item{rows}{XXX}

\item{mult}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

