% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_strata.R
\name{Combine.strata}
\alias{Combine.strata}
\alias{index.aggregate.combine}
\title{Sums mean, standard error, cv by aggregated area}
\usage{
Combine.strata(result, index)
}
\arguments{
\item{result}{Result}

\item{index}{Index}
}
\description{
Function used both in \code{\link{Calc.index}}.
}
\details{
location: /net/hafkaldi/export/u2/reikn/Splus5/SMB/GEOMETRY.NEW/.RData
}

