% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HAUSTRALLNewStrata_Stratifiering.R
\docType{data}
\name{haustralllist}
\alias{haustralllist}
\title{XXX}
\description{
XXX
location: /net/hafkaldi/export/u2/reikn/Splus5/HAUSTRALLNewStrata/Stratifiering/.RData
use_data(haustralllist)
}
\author{
Hoskuldur Bjornsson
}
