% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_reg_bcbreytt.R
\name{inside.reg.bcbreytt}
\alias{inside.reg.bcbreytt}
\title{Title}
\usage{
inside.reg.bcbreytt(data, only.sv.1to10 = F, ignore.latlon = F,
  ignore.area0 = T, option = 1)
}
\arguments{
\item{data}{xxx}

\item{only.sv.1to10}{xxx}

\item{ignore.latlon}{xxx}

\item{ignore.area0}{xxx}

\item{option}{xxx}
}
\description{
Title
}

