% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv_filter.R
\name{kv.filter}
\alias{kv.filter}
\title{XXX}
\usage{
kv.filter(kv, tegund, maxage, vikmork = pax::vikmork)
}
\arguments{
\item{kv}{XXX}

\item{tegund}{XXX}

\item{maxage}{XXX}

\item{vikmork}{XXX}
}
\description{
XXX

location: file:/net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

