% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAlk.R
\name{MakeAlk}
\alias{MakeAlk}
\title{XXX}
\usage{
MakeAlk(kvarnir, tegund, kyn = F, kynth = F, okynth.gildi = 1,
  kynth.gildi = c(2:4), lengd, aldur, Stodvar, FilterAldurLengd = T,
  NAOkynthAge = 3)
}
\arguments{
\item{kvarnir}{XXX}

\item{tegund}{XXX}

\item{kyn}{XXX}

\item{kynth}{XXX}

\item{okynth.gildi}{XXX}

\item{kynth.gildi}{XXX}

\item{lengd}{XXX}

\item{aldur}{XXX}

\item{Stodvar}{XXX}

\item{FilterAldurLengd}{XXX}

\item{NAOkynthAge}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

