% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLdist.R
\name{MakeLdist}
\alias{MakeLdist}
\title{XXX}
\usage{
MakeLdist(tegund, kyn = F, kynth = F, okynth.gildi = 1, kynth.gildi = 2,
  lengd, Stodvar, lengd.thyngd.data, lengdir, numer, talid = T, afli = NULL,
  std.toglengd = NULL, trollbreidd, area)
}
\arguments{
\item{tegund}{XXX}

\item{kyn}{XXX}

\item{kynth}{XXX}

\item{okynth.gildi}{XXX}

\item{kynth.gildi}{XXX}

\item{lengd}{XXX}

\item{Stodvar}{XXX}

\item{lengd.thyngd.data}{XXX}

\item{lengdir}{XXX}

\item{numer}{XXX}

\item{talid}{XXX}

\item{afli}{XXX}

\item{std.toglengd}{XXX}

\item{trollbreidd}{XXX}

\item{area}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

