% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLdistbyStation.R
\name{MakeLdistbyStation}
\alias{MakeLdistbyStation}
\title{Title}
\usage{
MakeLdistbyStation(lengdir, numer, tegund, kyn = F, kynth = F,
  okynth.gildi = 1, kynth.gildi = 2, lengd = seq(4.5, 130.5, by = 5),
  Stodvar, talid, std.toglengd = NULL, lengd.thyngd.data,
  stodvar.col = c("synis.id", "lat", "lon", "strata", "newstrata", "area",
  "toglengd"))
}
\arguments{
\item{lengdir}{XXX}

\item{numer}{XXX}

\item{tegund}{XXX}

\item{kyn}{XXX}

\item{kynth}{XXX}

\item{okynth.gildi}{XXX}

\item{kynth.gildi}{XXX}

\item{lengd}{XXX}

\item{Stodvar}{XXX}

\item{talid}{XXX}

\item{std.toglengd}{XXX}

\item{lengd.thyngd.data}{XXX}

\item{stodvar.col}{XXX}
}
\value{
XXX
}
\description{
location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

