% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_HAUSTRALLNewStrata_Stratifiering.R
\docType{data}
\name{newstratas_df}
\alias{newstratas_df}
\title{XXX}
\description{
XXX
x <- attributes(NEWSTRATAS)
newstratas_df <- data_frame(strata = 1:length(x$names),
                names = x$names,
                area = x$area,
                rall.area = x$rall.area,
                lon = x$pos[,2],
                lat = x$pos[,1])
use_data(newstratas)
}
\author{
Einar Hjorleifsson
}

