% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smooth_spline.R
\name{predict.smooth.spline}
\alias{predict.smooth.spline}
\title{XXX}
\usage{
\method{predict}{smooth.spline}(object, x, deriv = 0)
}
\arguments{
\item{object}{XXX}

\item{x}{XXX}

\item{deriv}{XXX}
}
\description{
XXX
}

