% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_alk1.R
\name{rbind.alk1}
\alias{rbind.alk1}
\title{XXX}
\usage{
rbind.alk1(fj1, fj2)
}
\arguments{
\item{fj1}{XXX}

\item{fj2}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

