% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reikna_thyngd.R
\name{Reikna.thyngd}
\alias{Reikna.thyngd}
\title{XXX}
\usage{
Reikna.thyngd(lengdir, l.th.gogn)
}
\arguments{
\item{lengdir}{XXX}

\item{l.th.gogn}{XXX}
}
\description{
XXX

location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData
}

