% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skala_med_afla.R
\name{Skala.med.afla}
\alias{Skala.med.afla}
\title{XXX}
\usage{
Skala.med.afla(lengdir, afli)
}
\arguments{
\item{lengdir}{XXX}

\item{afli}{XXX}
}
\description{
XXX

location "/net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData"
}

