% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skala_med_toglengd.R
\name{Skala.med.toglengd}
\alias{Skala.med.toglengd}
\title{XXX}
\usage{
Skala.med.toglengd(lengdir, Stodvar, std.toglengd)
}
\arguments{
\item{lengdir}{XXX}

\item{Stodvar}{XXX}

\item{std.toglengd}{XXX}
}
\description{
XXX

location "/net/hafkaldi/export/u2/reikn/Splus5/GETALK/.RData"
}

