% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skala_med_toldum.R
\name{Skala.med.toldum}
\alias{Skala.med.toldum}
\title{Skala.med.toldum}
\usage{
Skala.med.toldum(lengdir, numer, tegund)
}
\arguments{
\item{lengdir}{XXX}

\item{numer}{XXX}

\item{tegund}{XXX}
}
\description{
Skala.med.toldum
}
