% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GEOMETRYNEW.R
\docType{data}
\name{stratas_df}
\alias{stratas_df}
\title{XXX}
\description{
XXX

x <- attributes(STRATAS)
stratas_df <- data.frame(strata = as.integer(x$names),
area = x$area,
rall.area = x$rall.area,
lon = x$pos[,2],
lat = x$pos[,1])
use_data(stratas_df)
}
\author{
Einar Hjorleifsson
}

