% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GETALK_Weights.R
\docType{data}
\name{torskur.marsrall.wt.8592}
\alias{torskur.marsrall.wt.8592}
\title{torskur.marsrall.wt.8592}
\description{
XXX
location: /net/hafkaldi/export/u2/reikn/Splus5/GETALK/Weights/.RData
}
\author{
Hoskuldur Bjornsson
}
