% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_strata.R
\name{inside.strata}
\alias{inside.strata}
\title{inside.strata}
\usage{
inside.strata(stodvar, stratalist = ralllist, stratas = STRATAS)
}
\arguments{
\item{stodvar}{XXX}

\item{stratalist}{XXX}

\item{stratas}{XXX Defaults to the old strata (NOTE: Not in huskys original)}
}
\description{
location: /net/hafkaldi/export/u2/reikn/Splus5/SMB/GEOMETRY.NEW/.RData
}

